Refactor Kanban components to remove unnecessary containerRef prop and update imports for better structure
