+++ frontend/src/components/KanbanBoard/KanbanTaskbar/KanbanTaskbar.tsx
@@ -2,12 +2,12 @@ import { useState } from "react";
 import "./KanbanTaskbar.css";
 import KanbanCardStatus from "./KanbanCardStatus/KanbanCardStatus";
 import KanbanCardActions from "./KanbanCardActions/KanbanCardActions";
-import KanbanCardTabs from "./KanbanCardTabs/KanbanCardTabs";
 import type { BoardCard } from "../../../types/BoardCard";
 import type { ColumnHeaderBag } from "../../../types/ColumnHeaderBag";
 import type { TabType } from "../../../types/TabType";
 import { TabComponents } from "./TabComponents";
 import { assets } from "../../../assets/assets";
+import KanbanCardTabs from "./KanbanCardTabs/KanbanCardTabs.tsx";
 
 interface KanbanTaskbarProps {
   show: boolean;
