+++ frontend/src/types/ColumnOptionsBtn.ts
@@ -1,8 +1,6 @@
 import type { ColumnHeaderProps } from "./ColumnHeaderProps";
 import type { Dispatch, SetStateAction } from "react";
-import type { RefObject } from "react";
 
 export interface ColumnOptionsBtnProps extends ColumnHeaderProps {
   setIsEditable: Dispatch<SetStateAction<boolean>>;
-  containerRef: RefObject<HTMLDivElement | null>;
 }
